Move all formatting to a formatter class to remove duplication
